refactor(backend): rename graphql resolver files
